Added and Created Contents for Project Page
I worked on 2 specific Pages for the dashboard Section. The pages are dashboard_project.html and dashboard_projectstage.html, the 2 pages uses the dashboard-style.css for CSS formatting. 
The dashboard_project.html page comprises of an Accordion that bears the 5 different projects and gives users opportunity to read about the project and also select state and local government they wish to fund the project on.
The dashboard_projectstage.html shows the 5 different projects and information about their level of completion. It was designed using Tabs.
There are comments on the html and css page to guide on the flow of the design. 
Thanks.